refactor: decompose ReviewView into focused components (FE-16)
Extract the monolithic ReviewView.vue (1,659 lines) into focused
components and composables, each under 400 lines:

Components (in frontend/taskdeck-web/src/components/review/):
- ReviewHeader.vue (199 lines) - hero, board selector, action buttons
- ReviewSummaryCards.vue (74 lines) - statistics cards grid
- ReviewEmptyState.vue (47 lines) - empty state with guided actions
- ReviewProposalCard.vue (342 lines) - individual proposal card
- ReviewProposalActions.vue (229 lines) - approve/reject/apply/dismiss
- ReviewProposalDetails.vue (346 lines) - collapsible details sections

Composables (in frontend/taskdeck-web/src/composables/):
- useReviewProposals.ts (346 lines) - state, filtering, data loading
- useReviewActions.ts (159 lines) - proposal actions and diff toggle

ReviewView.vue is now 148 lines (orchestration + layout only).

All 45 existing ReviewView tests pass without modification.
Typecheck, build, and lint all pass with no new warnings.

Closes #856
