+++ frontend/taskdeck-web/src/components/review/ReviewProposalCard.vue
@@ -1,0 +1,342 @@
+<script setup lang="ts">
+import type { Proposal } from '../../types/automation'
+import {
+  normalizeProposalRiskLevel,
+  normalizeProposalSourceType,
+  normalizeProposalStatus,
+} from '../../utils/automation'
+import ReviewProposalActions from './ReviewProposalActions.vue'
+import ReviewProposalDetails from './ReviewProposalDetails.vue'
+
+const props = defineProps<{
+  proposal: Proposal
+  isExpired: boolean
+  isBusy: boolean
+  selectedDiffProposalId: string | null
+  selectedDiff: string | null
+  captureHref: string
+  proposalHref: string
+}>()
+
+defineEmits<{
+  (e: 'approve', proposalId: string): void
+  (e: 'reject', proposalId: string, riskLevel: Proposal['riskLevel']): void
+  (e: 'execute', proposalId: string): void
+  (e: 'toggle-diff', proposalId: string): void
+  (e: 'dismiss', proposalId: string): void
+  (e: 'open-board', boardId: string): void
+}>()
+
+function formatDate(value: string | null): string {
+  if (!value) {
+    return '-'
+  }
+  return new Date(value).toLocaleString()
+}
+
+function readableSummary(proposal: Proposal): string {
+  return proposal.presentation?.plainSummary || proposal.summary
+}
+
+function impactSummary(proposal: Proposal): string {
+  return proposal.presentation?.impactSummary
+    || `${proposal.operations.length} planned change${proposal.operations.length === 1 ? '' : 's'}.`
+}
+
+function getOperationHeadlines(proposal: Proposal): string[] {
+  if (proposal.presentation?.operationHeadlines?.length) {
+    return proposal.presentation.operationHeadlines
+  }
+  return proposal.operations.map((operation) => `${operation.actionType} ${operation.targetType}`)
+}
+
+function getAffectedEntities(proposal: Proposal) {
+  return proposal.presentation?.affectedEntities ?? []
+}
+
+function hasProvenanceContext(proposal: Proposal): boolean {
+  return !!captureSourceReference(proposal)
+}
+
+function captureSourceReference(proposal: Proposal): string | null {
+  if (normalizeProposalSourceType(proposal.sourceType) !== 'Queue') {
+    return null
+  }
+  if (!proposal.sourceReferenceId) {
+    return null
+  }
+  const trimmed = proposal.sourceReferenceId.trim()
+  return trimmed.length > 0 ? trimmed : null
+}
+
+function shortCorrelationId(correlationId: string): string {
+  const trimmed = correlationId.trim()
+  return trimmed.length > 8 ? trimmed.slice(0, 8) + '...' : trimmed
+}
+
+function reviewStatusClass(status: Proposal['status']): string {
+  if (props.isExpired) return 'td-review-status--expired'
+  const normalized = normalizeProposalStatus(status)
+  if (normalized === 'PendingReview') return 'td-review-status--pending'
+  if (normalized === 'Approved') return 'td-review-status--approved'
+  if (normalized === 'Expired') return 'td-review-status--expired'
+  if (normalized === 'Applied') return 'td-review-status--applied'
+  return 'td-review-status--secondary'
+}
+
+const statusLabels: Record<string, string> = {
+  PendingReview: 'Review required',
+  Approved: 'Approved, ready to apply',
+  Applied: 'Applied to board',
+  Rejected: 'Rejected',
+  Failed: 'Failed',
+  Expired: 'Expired',
+  Dismissed: 'Dismissed',
+}
+
+function reviewStatusLabel(status: Proposal['status']): string {
+  if (props.isExpired) return 'Expired'
+  const normalized = normalizeProposalStatus(status)
+  return statusLabels[normalized] ?? normalized
+}
+
+function riskLevelClass(riskLevel: Proposal['riskLevel']): string {
+  const normalized = normalizeProposalRiskLevel(riskLevel)
+  if (normalized === 'Low') return 'td-risk--low'
+  if (normalized === 'Medium') return 'td-risk--medium'
+  if (normalized === 'High') return 'td-risk--high'
+  if (normalized === 'Critical') return 'td-risk--critical'
+  return 'td-risk--low'
+}
+</script>
+
+<template>
+  <article
+    :id="`proposal-${proposal.id}`"
+    class="td-panel td-review-card"
+  >
+    <!-- Always visible: title, status badge, risk level, meta -->
+    <div class="td-review-card__header">
+      <div>
+        <h2 class="td-review-card__title">{{ readableSummary(proposal) }}</h2>
+        <div class="td-review-card__meta">
+          <span :class="['td-risk-badge', riskLevelClass(proposal.riskLevel)]">
+            {{ normalizeProposalRiskLevel(proposal.riskLevel) }} risk
+          </span>
+          <span>Created: {{ formatDate(proposal.createdAt) }}</span>
+          <span>Source: {{ normalizeProposalSourceType(proposal.sourceType) }}</span>
+        </div>
+      </div>
+      <span :class="['td-review-status', reviewStatusClass(proposal.status)]">
+        {{ reviewStatusLabel(proposal.status) }}
+      </span>
+    </div>
+
+    <!-- Impact cue always visible -->
+    <div class="td-review-card__presentation">
+      <span class="td-review-cue">{{ impactSummary(proposal) }}</span>
+    </div>
+
+    <!-- Action footer -->
+    <ReviewProposalActions
+      :proposal="proposal"
+      :is-expired="isExpired"
+      :is-busy="isBusy"
+      :selected-diff-proposal-id="selectedDiffProposalId"
+      @approve="$emit('approve', $event)"
+      @reject="(id, risk) => $emit('reject', id, risk)"
+      @execute="$emit('execute', $event)"
+      @toggle-diff="$emit('toggle-diff', $event)"
+      @dismiss="$emit('dismiss', $event)"
+    />
+
+    <!-- Collapsible details section -->
+    <ReviewProposalDetails
+      :proposal="proposal"
+      :operation-headlines="getOperationHeadlines(proposal)"
+      :affected-entities="getAffectedEntities(proposal)"
+      :has-provenance="hasProvenanceContext(proposal)"
+      :capture-href="captureHref"
+      :proposal-href="proposalHref"
+      :short-correlation-id="shortCorrelationId(proposal.correlationId)"
+      @open-board="$emit('open-board', $event)"
+    />
+
+    <div v-if="selectedDiffProposalId === proposal.id && selectedDiff" class="td-review-card__diff-wrapper">
+      <span class="td-review-card__diff-label">Operation details</span>
+      <pre class="td-review-card__diff" role="region" aria-label="Proposal operation diff">{{ selectedDiff }}</pre>
+    </div>
+  </article>
+</template>
+
+<style scoped>
+.td-review-card {
+  display: flex;
+  flex-direction: column;
+  gap: var(--td-space-3);
+  max-height: 70vh;
+  overflow-y: auto;
+}
+
+.td-review-card__header {
+  display: flex;
+  justify-content: space-between;
+  gap: var(--td-space-3);
+  align-items: flex-start;
+}
+
+.td-review-card__title {
+  margin: 0;
+  font-family: 'Manrope', system-ui, sans-serif;
+  font-size: var(--td-font-lg);
+  font-weight: 700;
+  color: var(--td-text-primary);
+}
+
+.td-review-card__meta {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--td-space-3);
+  margin-top: var(--td-space-1);
+  font-size: var(--td-font-xs);
+  color: var(--td-text-secondary);
+}
+
+.td-review-card__presentation {
+  display: flex;
+  flex-direction: column;
+  gap: var(--td-space-2);
+}
+
+.td-review-cue {
+  border-radius: var(--td-radius-pill, 999px);
+  background: var(--td-surface-container-high);
+  border: 1px solid var(--td-border-default);
+  color: var(--td-text-secondary);
+  font-size: var(--td-font-xs);
+  font-weight: 600;
+  padding: 0.25rem 0.625rem;
+}
+
+/* Risk level color-coded badges */
+.td-risk-badge {
+  border-radius: var(--td-radius-pill, 999px);
+  font-size: var(--td-font-xs);
+  font-weight: 700;
+  padding: 0.125rem 0.5rem;
+}
+
+.td-risk--low {
+  color: var(--td-color-success);
+  background: var(--td-color-success-light);
+}
+
+.td-risk--medium {
+  color: var(--td-color-warning);
+  background: var(--td-color-warning-light);
+}
+
+.td-risk--high {
+  color: var(--td-color-error);
+  background: var(--td-color-error-light);
+}
+
+.td-risk--critical {
+  color: var(--td-color-error);
+  background: var(--td-color-error-light);
+  border: 1px solid var(--td-color-error);
+}
+
+.td-review-status {
+  display: inline-flex;
+  align-items: center;
+  border-radius: var(--td-radius-pill, 999px);
+  border: 1px solid var(--td-border-default);
+  padding: 0.25rem 0.625rem;
+  font-size: var(--td-font-xs);
+  font-weight: 700;
+  white-space: nowrap;
+}
+
+.td-review-status--pending {
+  color: var(--td-color-warning);
+  background: var(--td-color-warning-light);
+  border-color: var(--td-color-warning);
+}
+
+.td-review-status--approved {
+  color: var(--td-color-success);
+  background: var(--td-color-success-light);
+  border-color: var(--td-color-success);
+}
+
+.td-review-status--applied {
+  color: var(--td-color-info);
+  background: var(--td-color-info-light);
+  border-color: var(--td-color-info);
+}
+
+.td-review-status--expired {
+  color: var(--td-color-warning);
+  background: var(--td-color-warning-light);
+  border-color: var(--td-color-warning);
+}
+
+.td-review-status--secondary {
+  color: var(--td-text-secondary);
+  background: var(--td-surface-container-high);
+}
+
+.td-review-card__diff-wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: var(--td-space-1);
+}
+
+.td-review-card__diff-label {
+  font-size: var(--td-font-xs);
+  color: var(--td-text-tertiary);
+  font-weight: 500;
+}
+
+.td-review-card__diff {
+  margin: 0;
+  padding: var(--td-space-3);
+  border-radius: var(--td-radius-md);
+  background: var(--td-surface-container-lowest);
+  color: var(--td-text-primary);
+  font-size: var(--td-font-xs);
+  overflow-x: auto;
+  border: 1px solid var(--td-border-ghost);
+  white-space: pre-wrap;
+  word-break: break-word;
+}
+
+@media (max-width: 900px) {
+  .td-review-card__header {
+    flex-direction: column;
+  }
+}
+
+@media (max-width: 640px) {
+  .td-review-card {
+    gap: var(--td-space-2);
+    padding: var(--td-space-4);
+    max-height: 80vh;
+  }
+
+  .td-review-card__title {
+    font-size: var(--td-font-base);
+  }
+
+  .td-review-card__meta {
+    flex-direction: column;
+    gap: var(--td-space-1);
+  }
+
+  .td-review-card__diff {
+    font-size: var(--td-font-xs);
+    padding: var(--td-space-2);
+  }
+}
+</style>
